refactor(auth): unify OAuth provider imports under oauthProviderIds
Updated all relevant files to import OAuthProvider from the new oauthProviderIds module, streamlining the codebase and avoiding circular dependencies. This change enhances maintainability and clarity in the authentication flow.
